# docs/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trackpatrol API",
    "description": "API doc Documentation for Trackpatrol",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "local server"
    },
    {
      "url": "https://trackpatrol.onrender.com/",
      "description": "production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/personnel/": {
      "post": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Create a new personnel",
        "description": "Create a new personnel (requires role : \"admin\" )",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personnel created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel created successfully response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel created successfully response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "sid already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Username already exists"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Username already exists"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/create personnel req.body"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Get all personnel with pagination",
        "description": "Get all personnel with pagination (requires role : \"admin\" )",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit per page",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personnel list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel list response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel list response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/personnel/bulk_create/get_template": {
      "get": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Get the csv template for bulk personnel creation",
        "description": "Endpoint to get the csv template for bulk personnel creation (requires role : \"admin\" )",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/admin/personnel/bulk_create": {
      "post": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Create multiple personnel records from a csv file",
        "description": "Endpoint to create multiple personnel records from a csv file (requires role : \"admin\") sample_file: https://docs.google.com/spreadsheets/d/1nREThXMnFuE5Rgo3XyY9hG5FYOPHe1J2o5NpZEsR26E/edit?usp=sharing",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personnel bulk created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel bulk creation successful response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel bulk creation successful response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "csv_file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "csv_file"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/personnel/search": {
      "get": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Search personnel by sid or name",
        "description": "Search personnel by sid or name (requires role : \"admin\" )",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit per page",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personnel list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel list response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel list response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/admin/personnel/bulk_delete": {
      "delete": {
        "tags": [
          "Admin : Personnel"
        ],
        "summary": "Delete multiple personnel by sid",
        "description": "Delete multiple personnel by sid. (requires role : \"admin\" )",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Personnel deleted in bulk response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel deleted in bulk response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel deleted in bulk response"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Personnel not found response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel not found response"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Personnel not found response"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Validation error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Internal server error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bulk delete personnel req.body"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}